mcp/requirements.txt
streamlit
uvicorn
httpx
python-dotenv
pandas
scikit-learn
mcp